refactor(workspace): share the file+backlinks read between GraphQL and REST (#177)
