+++ src/server/graphql/workspace.rs
@@ -6,8 +6,8 @@ use std::sync::Arc;
 use async_graphql::{ID, SimpleObject};
 
 use super::iso8601;
-use crate::company::extract_wikilinks;
 use crate::company::runtime::CompanyRuntime;
+use crate::company::workspace_links::file_with_backlinks;
 use crate::ports::workspace::{NodeKind, WorkspaceNode};
 
 /// One node (folder or file) in the workspace tree. Mirrors [`WorkspaceNode`].
@@ -58,11 +58,6 @@ pub struct WorkspaceFileGql {
     pub backlinks: Vec<FsNodeGql>,
 }
 
-/// The link target a file's name presents: its base name minus a `.md` suffix.
-fn link_target(name: &str) -> &str {
-    name.strip_suffix(".md").unwrap_or(name)
-}
-
 /// Resolves `Company.workspaceTree`.
 pub(crate) async fn resolve_tree(
     runtime: &Arc<CompanyRuntime>,
@@ -72,36 +67,26 @@ pub(crate) async fn resolve_tree(
 }
 
 /// Resolves `Company.workspaceFile(id)`, returning null when absent.
+///
+/// The node + content + backlink scan is
+/// [`file_with_backlinks`](crate::company::workspace_links::file_with_backlinks),
+/// shared with the REST `GET …/workspace/file/{id}` route the console reads, so
+/// the two surfaces can never report different backlinks for the same note.
 pub(crate) async fn resolve_file(
     runtime: &Arc<CompanyRuntime>,
     id: &str,
 ) -> async_graphql::Result<Option<WorkspaceFileGql>> {
-    let Some((node, content)) = runtime.workspace().read(runtime.id(), id).await? else {
+    let Some((node, content, backlinks)) =
+        file_with_backlinks(runtime.workspace().as_ref(), runtime.id(), id).await?
+    else {
         return Ok(None);
     };
-    let target = link_target(&node.name).to_string();
-
-    // Backlinks: scan every other file node's content for a `[[target]]` link.
-    let mut backlinks = Vec::new();
-    for other in runtime.workspace().tree(runtime.id()).await? {
-        if other.id == node.id || !matches!(other.kind, NodeKind::File) {
-            continue;
-        }
-        if let Some((other_node, other_content)) =
-            runtime.workspace().read(runtime.id(), &other.id).await?
-            && extract_wikilinks(&other_content)
-                .iter()
-                .any(|link| link == &target)
-        {
-            backlinks.push(FsNodeGql::from(other_node));
-        }
-    }
 
     Ok(Some(WorkspaceFileGql {
         id: ID(node.id),
         name: node.name,
         content,
         updated_at: iso8601(node.updated_at_millis),
-        backlinks,
+        backlinks: backlinks.into_iter().map(FsNodeGql::from).collect(),
     }))
 }
